卡密列表增加筛选

+++ WangShunManager/Dtos/PreCardRowDto.cs
@@ -2,8 +2,8 @@
 {
     public class PreCardRowDto
     {
-        public int Id { get; set; }
-        public int BatchId { get; set; }
+        public long Id { get; set; }
+        public long BatchId { get; set; }
         public string CardId { get; set; }
         public int CardState { get; set; }
         public int ManagedState { get; set; }
